Introduce new Task class to deal with non-uniform tasks
The API returns two different Task.state one a string, one an object.
