+++ src/main/java/mesosphere/marathon/client/model/v2/TaskCondition.java
@@ -1,0 +1,13 @@
+package mesosphere.marathon.client.model.v2;
+
+public class TaskCondition {
+    private String str;
+
+    public String getStr() {
+        return str;
+    }
+
+    public void setStr(String str) {
+        this.str = str;
+    }
+}
